(FACT-1082) Fix loading custom fact with win32ole
Loading a custom fact on the command-line that uses win32ole results in an
error, because Ruby's win32ole library initializes COM with
single-threaded and Facter previously initialized it with multi-threaded.

Change Facter to initialize COM single-threaded, and add tests verifying
custom facts can use win32ole successfully.
